+++ bcbio/rnaseq/kallisto.py
@@ -15,6 +15,55 @@ from bcbio.pipeline import config_utils
 from bcbio.rnaseq import umi
 from bcbio.bam import fasta
 
+def run_kallisto_rnaseq(data):
+    samplename = dd.get_sample_name(data)
+    work_dir = dd.get_work_dir(data)
+    kallisto_dir = os.path.join(work_dir, "kallisto", samplename)
+    gtf_file = dd.get_gtf_file(data)
+    files = dd.get_input_sequence_files(data)
+    if len(files) == 2:
+        fq1, fq2 = files
+    else:
+        fq1, fq2 = files[0], None
+    assert file_exists(gtf_file), "%s was not found, exiting." % gtf_file
+    fasta_file = dd.get_ref_file(data)
+    assert file_exists(fasta_file), "%s was not found, exiting." % fasta_file
+    assert fq2, ("bcbio doesn't support kallisto for single-end reads, we can "
+                 "add support for this if you open up an issue about it here: "
+                 "https://github.com/chapmanb/bcbio-nextgen/issues")
+    out_file = kallisto_rnaseq(fq1, fq2, kallisto_dir, gtf_file, fasta_file, data)
+    data = dd.set_kallisto_quant(data, out_file)
+    return [[data]]
+
+def kallisto_rnaseq(fq1, fq2, kallisto_dir, gtf_file, fasta_file, data):
+    samplename = dd.get_sample_name(data)
+    quant_dir = os.path.join(kallisto_dir, "quant")
+    safe_makedir(kallisto_dir)
+    num_cores = dd.get_num_cores(data)
+    strandedness = dd.get_strandedness(data).lower()
+    kallisto = config_utils.get_program("kallisto", dd.get_config(data))
+    index = kallisto_index(gtf_file, fasta_file, data, kallisto_dir)
+    fusion_flag = "--fusion" if dd.get_fusion_mode(data) else ""
+    single_flag = "--single" if not fq2 else ""
+    fraglength_flag = "--fragment-length=200" if not fq2 else ""
+    sd_flag = "--sd=25" if not fq2 else ""
+    bootstrap_flag = "--bootstrap-samples=30"
+    fq2 = "" if not fq2 else fq2
+    if not fq2:
+        logger.warning("kallisto was run on single-end data and we set the "
+          "estimated fragment length to 200 and the standard "
+          "deviation to 25, if these don't reflect your data then "
+          "the results may be inaccurate. Use with caution. See "
+          "https://groups.google.com/forum/#!topic/kallisto-sleuth-users/h5LeAlWS33w "
+          "for details.")
+    cmd = ("{kallisto} quant {fusion_flag} -t {num_cores} {single_flag} "
+           "{fraglength_flag} {sd_flag} {bootstrap_flag} "
+           "-o {tx_out_dir} -i {index} {fq1} {fq2}")
+    with file_transaction(data, quant_dir) as tx_out_dir:
+        message = ("Quantifying transcripts with kallisto.")
+        do.run(cmd.format(**locals()), message, None)
+    return quant_dir
+
 def run_kallisto_singlecell(data):
     samplename = dd.get_sample_name(data)
     work_dir = dd.get_work_dir(data)
@@ -56,7 +105,7 @@ def kallisto_index(gtf_file, ref_file, data, out_dir):
     out_dir = os.path.join(out_dir, "index")
     out_stem = dd.get_genome_build(data)
     if dd.get_disambiguate(data):
-        out_fname = "-".join([out_fname] + dd.get_disambguate(data))
+        out_stem = "-".join([out_stem] + dd.get_disambiguate(data))
     index_dir = os.path.join(out_dir, out_stem)
     kallisto = config_utils.get_program("kallisto", dd.get_config(data))
     if dd.get_transcriptome_fasta(data):